+++ .gitignore
@@ -1,3 +1,3 @@
-my_venv
+venv
 __pycache__
 .env
