+++ app/__init__.py
@@ -4,13 +4,34 @@ from flask_login import LoginManager
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
 
-app = Flask(__name__)
-app.config.from_object(Config)
+
 
 login_manager = LoginManager()
-login_manager.init_app(app)
+db = SQLAlchemy()
+migrate = Migrate()
 
-db = SQLAlchemy(app)
-migrate = Migrate(app, db)
 
-from app import routes, models
+def create_app():
+
+    app = Flask(__name__)
+    app.config.from_object(Config)
+
+    #register packages
+    login_manager.init_app(app)
+    db.init_app(app)
+    migrate.init_app(app, db)
+
+    #Importing Blueprint
+    from app.blueprints.main import main
+    from app.blueprints.auth import auth
+
+    #Register Blueprint
+    app.register_blueprint(main)
+    app.register_blueprint(auth)
+
+    return app
+
+
+
+
+# from app import routes, models
