+++ app/templates/base.html
@@ -6,6 +6,12 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
     <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css')}}">
+
+    <link rel="apple-touch-icon" sizes="180x180" href="{{ url_for('static', filename='images/apple-touch-icon.png')}}">
+    <link rel="icon" type="image/png" sizes="32x32" href="{{ url_for('static', filename='images/favicon-32x32.png')}}">
+    <link rel="icon" type="image/png" sizes="16x16" href="{{ url_for('static', filename='images/favicon-16x16.png')}}">
+    <link rel="manifest" href="{{ url_for('static', filename='images/site.webmanifest')}}">
+,
     <title>{% block title %} {% endblock %}</title>
 </head>
 <body>
@@ -19,24 +25,24 @@
             <ul class="navbar-nav">
               {% if current_user.is_authenticated %}
                 <li class="nav-item">
-                  <a class="nav-link active" aria-current="page" href="{{ url_for('home') }}">Home</a>
+                  <a class="nav-link active" aria-current="page" href="{{ url_for('main.home') }}">Home</a>
                 </li>
                 <li class="nav-item">
-                  <a class="nav-link" href="{{ url_for('ergast') }}">Ergast</a>
+                  <a class="nav-link" href="{{ url_for('main.ergast') }}">Ergast</a>
                 </li>
                 <li class="nav-item">
-                  <a class="nav-link" href="{{ url_for('students') }}">Students</a>
+                  <a class="nav-link" href="{{ url_for('main.students') }}">Students</a>
                 </li>
                 <li class="nav-item">
                   <p class="nav-link">Hello, {{ current_user.first_name }}</>
                 </li>
-                <a class="btn btn-secondary" href="{{ url_for('logout')}}">Logout</a>
+                <a class="btn btn-secondary" href="{{ url_for('auth.logout')}}">Logout</a>
               {% else %}
                 <li class="nav-item">
-                  <a class="nav-link" href="{{ url_for('login') }}">Login</a>
+                  <a class="nav-link" href="{{ url_for('auth.login') }}">Login</a>
                 </li>
                 <li class="nav-item">
-                  <a class="nav-link" href="{{ url_for('signup') }}">Signup</a>
+                  <a class="nav-link" href="{{ url_for('auth.signup') }}">Signup</a>
                 </li>
               {% endif %}
             </ul>
